Cleaning up classes
