+++ ena_objects/characteristic.py
@@ -15,74 +15,6 @@ class IsaBase:
         [validate_dict(dict=dict, key=key) for key in mandatory_keys]
 
 
-# class Category(IsaBase):
-#     """
-#     This represents a category object in a Characteristic
-#     """
-
-#     def __init__(self, id: str, name: str) -> None:
-#         self.id = id
-#         self.name = name
-
-#     mandatory_keys = ["@id"]
-
-#     @classmethod
-#     def from_dict(self, dict: Dict, categories: Dict):
-#         super().check_dict_keys(dict, self.mandatory_keys)
-
-#         return self(id=dict["@id"], name=fetch_category_name(categories, dict))
-
-
-# class Value(IsaBase):
-#     """
-#     This represents a Value object in a Characteristic
-#     """
-
-#     mandatory_keys = ["annotationValue", "termSource", "termAccession"]
-
-#     def __init__(
-#         self, annotation_value: str, term_source: str = "", term_accession: str = ""
-#     ) -> None:
-#         self.annotation_value = annotation_value
-#         self.term_source = term_source
-#         self.term_accession = term_accession
-
-#     @classmethod
-#     def from_dict(self, dict: Dict):
-#         super().check_dict_keys(dict, self.mandatory_keys)
-
-#         return Value(
-#             annotation_value=dict["annotationValue"],
-#             term_accession=dict["termAccession"],
-#             term_source=dict["termSource"],
-#         )
-
-
-# class Unit(IsaBase):
-#     """
-#     This represents the Unit object in a Characteristic
-#     """
-
-#     mandatory_keys = ["termSource", "termAccession", "comments"]
-
-#     def __init__(
-#         self, term_source: str, term_accession: str, comments: List[str]
-#     ) -> None:
-#         self.term_source = term_source
-#         self.term_accession = term_accession
-#         self.comments = comments
-
-#     @classmethod
-#     def from_dict(self, dict: Dict):
-#         super().check_dict_keys(dict, self.mandatory_keys)
-
-#         return Unit(
-#             term_source=dict["termSource"],
-#             term_accession=dict["termAccession"],
-#             comments=dict["comments"],
-#         )
-
-
 def fetch_category_name(categories: Dict, name: str) -> str:
     for cat in categories:
         if name["@id"] == cat["id"]:
@@ -123,3 +55,36 @@ class Characteristic(IsaBase):
             "category": self.category,
             "value": self.value,
         }
+
+
+class OtherMaterialCharacteristic(Characteristic):
+    """
+    This class represents the other material object.
+    """
+
+    def __init__(self, category: Dict, value: str) -> None:
+        super().__init__(category, value)
+
+    @classmethod
+    def from_dict(cls, dict: Dict, characteristics_categories: Dict):
+        return super().from_dict(dict, characteristics_categories)
+
+    def to_dict(self) -> Dict:
+        return super().to_dict()
+
+
+class ParameterValue(Characteristic):
+    """
+    This class represents a paramenter value in the isa study
+    and is inherited from the Characteristic class
+    """
+
+    def __init__(self, category: Dict, value: str) -> None:
+        super().__init__(category, value)
+
+    @classmethod
+    def from_dict(self, dict: Dict, parameters: Dict):
+        return super().from_dict(dict, parameters)
+
+    def to_dict(self) -> Dict:
+        return super().to_dict()
